+++ src/data/mocks/disclaimer.js
@@ -2,9 +2,8 @@ const mockDisclaimerData = `This website is a SFoxInsurance engine Demo version.
 This website uses cookies and local storage to save your data, you can at any time reset 
 this data by using Settings button (click two times on start application) and Reset Settings.
 Username: player, pwd: ads123 to show Account Details.
-SFoxInsurance does not provide a credit advisory service for the purposes of the Insurance Intermediates Act.
-The borrower makes the decision of taking out a loan, who accesses the suitability of the loan product and contractual terms to his/her personal
-loan interest, need and financial situation on the basis of the information and warnings presented by the bank
-and is responsible for the consequences related to concluding the agreement.`;
+SFoxInsurance does not provide warranty and purpose of the website only to demonstrate abilities 
+of Sfox Engine. Assets, image are not owned by Author of this engine, there are property of respective owners.
+Please respect engine's current BSD-3 License.`;
 
 export default mockDisclaimerData;