+++ src/pages/additional/account/account.js
@@ -1,6 +1,9 @@
 // @ts-nocheck
-import { LoggerService } from '../../../services';
-import { CustomEvents } from '../../../settings';
+import { SaveObjects } from '../../../components/common/saves';
+import { LoggerService, StyleService } from '../../../services';
+import { objectPropertyAmount } from '../../../services/utils';
+import DataStorage from '../../../services/storage';
+import { CommonEvents, CustomEvents } from '../../../settings';
 
 class AccountPage extends HTMLElement {
     constructor() {
@@ -8,6 +11,17 @@ class AccountPage extends HTMLElement {
       this.shadow = this.attachShadow({ mode: 'closed' });
       this.isAccVisible = false;
       this.loggedUser = {};
+      this.logoutSet = false;
+      this.statuses = {
+        offline: 'offline',
+        loggedIn: 'logged-in'
+      };
+      this.status = this.statuses.offline;
+      this.events = {
+        init: 'init',
+        login: 'login'
+      };
+      this.storage = new DataStorage();
     }
   
     connectedCallback() {
@@ -16,32 +30,88 @@ class AccountPage extends HTMLElement {
     }
 
     initForm() {
-      this.showUserDetails();
+      this.getSaveAccount();
+      this.showUserDetails(this.loggedUser ? this.events.login : this.events.init);
+      this.$logout = this.shadow.getElementById('logout');
+      this.setLogoutHandler();
       document.addEventListener(CustomEvents.users.login, (evt) => {
         if (!evt.detail || !evt.detail.value) {
           LoggerService.warn('Login data missing!');
           return;
         }
         this.loggedUser = evt.detail.value;
-        this.setAvailable(true);
+        this.setAvailable(true, this.events.login);
       });
     }
 
-    setAvailable(accessible) {
-      this.isAccVisible = accessible;
-      if (accessible) {
-        this.showUserDetails();
+    disconnectedCallback() {
+      this.$logout.removeEventListener(CommonEvents.click, null);
+      document.removeEventListener(CustomEvents.users.login, null);
+    }
+
+    setLogoutHandler() {
+      this.$logout = this.shadow.getElementById('logout');
+      if (!this.logoutSet && this.$logout) {
+        this.$logout.addEventListener(CommonEvents.click, () => {
+          this.logout();
+        });
+        this.toggleLogout(true);
       }
     }
 
-    showUserDetails() {
-      if (Object.keys(this.loggedUser).length < 1) {
+    getSaveAccount() {
+      const saved = this.storage.getObject(SaveObjects.account);
+      if (saved && saved.status === this.statuses.loggedIn) {
+        this.loggedUser = saved;
+        this.setStatus(this.statuses.loggedIn);
+        this.toggleLogin(false);
+      }
+    }
+
+    setAvailable(accessible, evt) {
+      this.isAccVisible = accessible;
+      if (accessible) {
+        this.toggleLogin(false);
+        this.showUserDetails(evt);
+        this.setStatus(this.statuses.loggedIn);
+        this.setLogoutHandler();
+      }
+    }
+
+    setStatus(toggle) {
+      this.status = toggle;
+    }
+
+    toggleLogout(toggle){ 
+      this.logoutSet = toggle;
+    }
+
+    logout() {
+      this.setStatus(this.statuses.offline);
+      this.setDetails('');
+      this.toggleLogin(true);
+      this.toggleLogout(false);
+      this.saveObjectAndStatus(this.statuses.offline);
+      this.$logout.removeEventListener(CommonEvents.click, null);
+    }
+
+    showUserDetails(event = this.events.init) {
+      if (objectPropertyAmount(this.loggedUser) < 1) {
         return;
       }
 
       const { username, email, name, surname } = this.loggedUser;
+      if (event === this.events.login) {
+        this.saveObjectAndStatus(this.statuses.loggedIn);
+      }
+
       const html = `
-        <div>
+        <div class="details">
+          <div>
+            <action-button id="logout" label="Logout" type="action"></action-button>
+          </div>
+
+          <h3>Account details</h3>
           <p> username: <b>${username}</b> </p>
           <p> email: <b>${email}</b> </p>
           <p> name: <b>${name}</b> <p>
@@ -49,28 +119,47 @@ class AccountPage extends HTMLElement {
         </div>
       `;
 
+      this.setDetails(html);
+    }
+
+    setDetails(html) {
       const el = this.shadow.getElementById('userDetails');
       el.innerHTML = html;
     }
+
+    toggleLogin(visible) {
+      const $login = this.shadow.getElementById('login');
+      StyleService.setDisplay($login, visible);
+    }
   
+    saveObjectAndStatus(status) {
+      this.loggedUser['status'] = status;
+      this.storage.saveObject(SaveObjects.account, this.loggedUser);
+    }
+
     render() {
       this.shadow.innerHTML = `
           <style>
-              .account {
-                display: flex;
-                flex-direction: column;
-                align-items: center;
-                justify-content: center;
-                padding: 20px;
+            .account {
+              display: flex;
+              flex-direction: column;
+              align-items: center;
+              justify-content: center;
+              padding: 20px;
+            }
+
+            .details {
+              & div:nth-child(1) {
+                padding-left: 100px;
+              }
             }
           </style>
           <form>
             <div class="account">
-              <section>
+              <section id="login">
                 <account-login></account-login>
               </section>
               <section>
-                <h3>Account details</h3>
                 <div id="userDetails"></div>
               </section>
             </div>