+++ src/services/utils/index.js
@@ -3,3 +3,4 @@ export { showComponent } from './render';
 export { getVersionFromPackage } from './strings';
 export { randomInteger, sample } from './number';
 export { isMobile } from './mobile';
+export { objectPropertyAmount } from './object';