+++ src/services/utils/object.js
@@ -1,0 +1,3 @@
+export const objectPropertyAmount = (obj) => {
+    return obj ? Object.keys(obj).length : 0;
+};